Conflicts:
	src/Bpp/Phyl.all
	src/Bpp/Phyl/App/PhylogeneticsApplicationTools.cpp
	src/Bpp/Phyl/App/PhylogeneticsApplicationTools.h
	src/Bpp/Phyl/Io.all
	src/Bpp/Phyl/Io/BppOFrequenciesSetFormat.cpp
	src/Bpp/Phyl/Io/BppOFrequenciesSetFormat.h
	src/Bpp/Phyl/Io/BppOSubstitutionModelFormat.cpp
	src/Bpp/Phyl/Io/BppOSubstitutionModelFormat.h
	src/Bpp/Phyl/Io/IoFrequenciesSet.h
	src/Bpp/Phyl/Io/IoFrequenciesSetFactory.cpp
	src/Bpp/Phyl/Io/IoFrequenciesSetFactory.h
	src/Bpp/Phyl/Io/IoSubstitutionModel.h
	src/Bpp/Phyl/Io/IoSubstitutionModelFactory.cpp
	src/Bpp/Phyl/Io/IoSubstitutionModelFactory.h
	src/Bpp/Phyl/Mapping/SubstitutionRegister.h
	src/Bpp/Phyl/Model/AbstractBiblioSubstitutionModel.cpp
	src/Bpp/Phyl/Model/AbstractSubstitutionModel.cpp
	src/Bpp/Phyl/Model/BinarySubstitutionModel.cpp
	src/Bpp/Phyl/Model/Codon/AbstractCodonDistanceSubstitutionModel.cpp
	src/Bpp/Phyl/Model/Codon/AbstractCodonDistanceSubstitutionModel.h
	src/Bpp/Phyl/Model/Codon/AbstractCodonFrequenciesSubstitutionModel.h
	src/Bpp/Phyl/Model/Codon/AbstractCodonPhaseFrequenciesSubstitutionModel.cpp
	src/Bpp/Phyl/Model/Codon/AbstractCodonPhaseFrequenciesSubstitutionModel.h
	src/Bpp/Phyl/Model/Codon/AbstractCodonSubstitutionModel.cpp
	src/Bpp/Phyl/Model/Codon/AbstractCodonSubstitutionModel.h
	src/Bpp/Phyl/Model/Codon/GY94.h
	src/Bpp/Phyl/Model/Codon/YN98.cpp
	src/Bpp/Phyl/Model/Codon/YNGKP_M1.cpp
	src/Bpp/Phyl/Model/Codon/YNGKP_M2.cpp
	src/Bpp/Phyl/Model/Codon/YNGKP_M3.cpp
	src/Bpp/Phyl/Model/Codon/YNGKP_M8.cpp
	src/Bpp/Phyl/Model/FrequenciesSet/CodonFrequenciesSet.cpp
	src/Bpp/Phyl/Model/FrequenciesSet/CodonFrequenciesSet.h
	src/Bpp/Phyl/Model/FrequenciesSet/FrequenciesSet.cpp
	src/Bpp/Phyl/Model/FrequenciesSet/NucleotideFrequenciesSet.h
	src/Bpp/Phyl/Model/FrequenciesSet/ProteinFrequenciesSet.h
	src/Bpp/Phyl/Model/FrequenciesSet/WordFrequenciesSet.cpp
	src/Bpp/Phyl/Model/G2001.h
	src/Bpp/Phyl/Model/MixtureOfASubstitutionModel.h
	src/Bpp/Phyl/Model/MixtureOfSubstitutionModels.cpp
	src/Bpp/Phyl/Model/Nucleotide/L95.cpp
	src/Bpp/Phyl/Model/Nucleotide/YpR.h
	src/Bpp/Phyl/Model/Protein/DSO78.h
	src/Bpp/Phyl/Model/Protein/JCprot.h
	src/Bpp/Phyl/Model/Protein/LLG08_EHO.cpp
	src/Bpp/Phyl/Model/Protein/LLG08_EHO.h
	src/Bpp/Phyl/Model/Protein/LLG08_EX2.cpp
	src/Bpp/Phyl/Model/Protein/LLG08_EX2.h
	src/Bpp/Phyl/Model/Protein/LLG08_EX3.cpp
	src/Bpp/Phyl/Model/Protein/LLG08_EX3.h
	src/Bpp/Phyl/Model/Protein/LLG08_UL2.cpp
	src/Bpp/Phyl/Model/Protein/LLG08_UL2.h
	src/Bpp/Phyl/Model/Protein/LLG08_UL3.cpp
	src/Bpp/Phyl/Model/Protein/LLG08_UL3.h
	src/Bpp/Phyl/Model/SubstitutionModelSet.cpp
	src/Bpp/Phyl/Model/WordSubstitutionModel.cpp
	src/Bpp/Phyl/OptimizationTools.cpp

+++ src/Bpp/Phyl/BipartitionList.h
@@ -129,23 +129,23 @@ class BipartitionList:
 
   public:
 
-    unsigned int getNumberOfElements() const { return elements_.size(); }
+    size_t getNumberOfElements() const { return elements_.size(); }
 
     const std::vector<std::string>& getElementNames() const { return elements_; }
 
-    unsigned int getNumberOfBipartitions() const { return bitBipartitionList_.size(); }
+    size_t getNumberOfBipartitions() const { return bitBipartitionList_.size(); }
 
     const std::vector<int*> & getBitBipartitionList() const { return bitBipartitionList_; }
 
-    std::map<std::string, bool> getBipartition(unsigned int i) const throw (Exception);
+    std::map<std::string, bool> getBipartition(size_t i) const throw (Exception);
 
-    int* getBitBipartition(unsigned int i) throw (Exception);
+    int* getBitBipartition(size_t i) throw (Exception);
 
     bool haveSameElementsThan(std::map<std::string, bool>& bipart) const;
 
     void addBipartition(std::map<std::string, bool>& bipart, bool checkElements = 1) throw(Exception);
 
-    void deleteBipartition(unsigned int i) throw(Exception);
+    void deleteBipartition(size_t i) throw(Exception);
 
     bool isSorted() const { return sorted_; }
 
@@ -153,7 +153,7 @@ class BipartitionList:
 
     bool containsBipartition(std::map<std::string, bool>& bipart, bool checkElements = 1) const throw(Exception);
 
-    bool areIdentical(unsigned int k1, unsigned int k2) const throw(Exception);
+    bool areIdentical(size_t k1, size_t k2) const throw(Exception);
 
     void removeRedundantBipartitions();
 
@@ -165,7 +165,7 @@ class BipartitionList:
      * and B1 contains A2) or (B1 contains A1 and A2 contains B2) or (B2 contains A1 and A2
      * contains B1). Only compatible bipartitions can belong to the same tree.
      */
-    bool areCompatible(unsigned int k1, unsigned int k2) const throw(Exception);
+    bool areCompatible(size_t k1, size_t k2) const throw(Exception);
 
     /**
      * @brief Tells whether all bipartitions in the list are compatible with each other
@@ -194,12 +194,12 @@ class BipartitionList:
     /**
      * @brief Replaces ones by zeros and zeros by ones in the ith bipartition
      */
-    void flip(unsigned int i) throw(Exception);
+    void flip(size_t i) throw(Exception);
 
     /**
      * @brief Returns the size of the smallest of the two partitions (e.g. 1 for external branches)
      */
-    unsigned int getPartitionSize(unsigned int i) const throw(Exception);
+    size_t getPartitionSize(size_t i) const throw(Exception);
 
     /**
      * @brief Sort bipartitions by partition size
@@ -224,7 +224,7 @@ class BipartitionList:
 
   private:
 
-    std::vector<std::string> buildBitBipartitions(const Node* nd, std::vector<int*>& bitbip, const std::vector<std::string> & elements, unsigned int* cpt, std::vector<int>* index) const;
+    std::vector<std::string> buildBitBipartitions(const Node* nd, std::vector<int*>& bitbip, const std::vector<std::string> & elements, size_t* cpt, std::vector<int>* index) const;
 
 };
