+++ src/Bpp/Phyl/Likelihood/AbstractTreeLikelihood.cpp
@@ -46,7 +46,7 @@ using namespace bpp;
 Vdouble AbstractTreeLikelihood::getLikelihoodForEachSite() const
 {
 	Vdouble l(getNumberOfSites());
-	for(unsigned int i = 0; i < l.size(); i++) l[i] = getLikelihoodForASite(i);
+	for (size_t i = 0; i < l.size(); i++) l[i] = getLikelihoodForASite(i);
 	return l;
 }
 
@@ -55,7 +55,7 @@ Vdouble AbstractTreeLikelihood::getLikelihoodForEachSite() const
 Vdouble AbstractTreeLikelihood::getLogLikelihoodForEachSite() const
 {
 	Vdouble l(getNumberOfSites());
-	for(unsigned int i = 0; i < l.size(); i++) l[i] = getLogLikelihoodForASite(i);
+	for (size_t i = 0; i < l.size(); i++) l[i] = getLogLikelihoodForASite(i);
 	return l;
 }
 
@@ -64,13 +64,13 @@ Vdouble AbstractTreeLikelihood::getLogLikelihoodForEachSite() const
 VVdouble AbstractTreeLikelihood::getLikelihoodForEachSiteForEachState() const
 {
 	VVdouble l(getNumberOfSites());
-	for(unsigned int i = 0; i < l.size(); i++)
+	for (size_t i = 0; i < l.size(); i++)
   {
-		Vdouble * l_i = & l[i];
+		Vdouble* l_i = & l[i];
 		l_i->resize(getNumberOfStates());
-		for(unsigned int x = 0; x < l_i->size(); x++)
+		for (size_t x = 0; x < l_i->size(); x++)
     {
-			(* l_i)[x] = getLikelihoodForASiteForAState(i, x);
+			(* l_i)[x] = getLikelihoodForASiteForAState(i, static_cast<int>(x));
 		}
 	}
 	return l;
@@ -81,13 +81,13 @@ VVdouble AbstractTreeLikelihood::getLikelihoodForEachSiteForEachState() const
 VVdouble AbstractTreeLikelihood::getLogLikelihoodForEachSiteForEachState() const
 {
 	VVdouble l(getNumberOfSites());
-	for (unsigned int i = 0; i < l.size(); i++)
+	for (size_t i = 0; i < l.size(); i++)
   {
-		Vdouble * l_i = & l[i];
+		Vdouble* l_i = & l[i];
 		l_i->resize(getNumberOfStates());
-		for(unsigned int x = 0; x < l_i->size(); x++)
+		for (size_t x = 0; x < l_i->size(); x++)
     {
-			(* l_i)[x] = getLogLikelihoodForASiteForAState(i, x);
+			(* l_i)[x] = getLogLikelihoodForASiteForAState(i, static_cast<int>(x));
 		}
 	}
 	return l;
